// core/packages/web-client/src/app/scenarios/store/types.ts
import { ScenarioJSON } from "ditto-shared";

export const enum ScenariosActionTypes {
  SCENARIO_CREATION_SUCCEEDED = "SCENARIO_CREATION_SUCCEEDED",
  SCENARIO_CREATION_STARTED = "SCENARIO_CREATION_STARTED",
  SCENARIO_CREATION_FAILED = "SCENARIO_CREATION_FAILED",
  RECIEVE_SCENARIOS = "RECIEVE_SCENARIOS",
  SCENARIO_PLAY_FAILED = "SCENARIO_PLAY_FAILED",
  SCENARIO_PLAY_SUCCEEDED = "SCENARIO_PLAY_SUCCEEDED",
  RECIEVE_SCENARIO = "RECIEVE_SCENARIO",
  SCENARIO_UPDATE_STARTED = "SCENARIO_UPDATE_STARTED",
  SCENARIO_UPDATE_SUCCEEDED = "SCENARIO_UPDATE_SUCCEEDED",
  SCENARIO_UPDATE_FAILED = "SCENARIO_UPDATE_FAILED",
  SCENARIO_DELETE_STARTED = "SCENARIO_DELETE_STARTED",
  SCENARIO_DELETE_SUCCEEDED = "SCENARIO_DELETE_SUCCEEDED",
  SCENARIO_DELETE_FAILED = "SCENARIO_DELETE_FAILED"
}

export interface ScenariosState{
    readonly all : ScenarioJSON[];
    readonly selected : ScenarioJSON | undefined;
}